+++ gaios/UserSettings/Controllers/UserSettingsViewController.swift
@@ -415,6 +415,7 @@ extension UserSettingsViewController: UITableViewDelegate, UITableViewDataSource
             let storyboard = UIStoryboard(name: "UserSettings", bundle: nil)
             if let vc = storyboard.instantiateViewController(withIdentifier: "TwoFactorAuthenticationViewController") as? TwoFactorAuthenticationViewController {
                 navigationController?.pushViewController(vc, animated: true)
+                vc.delegate = self
             }
 
         case .Pgp:
@@ -686,5 +687,10 @@ extension UserSettingsViewController {
         DispatchQueue.main.async {
             self.present(alert, animated: true, completion: nil)
         }
+    }
+}
+extension UserSettingsViewController: TwoFactorAuthenticationViewControllerDelegate {
+    func userLogout() {
+        self.logout()
     }
 }