+++ CHANGELOG.md
@@ -13,8 +13,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 - Show TEST instead of BTC for a testnet wallet
+- Link settings and 2fa data into session to better handle 2fa reset/cancel
 
 ### Fixed
+
 
 ## [3.7.3] - 2021-09-30